Created the StockManagerSingleton class
added the class to a new package group.hw3.stockManager. In the class I created variables to hold the file path and the singular stockManagerSingleton object. Also wrote all of the prototype methods for the StockManagerSingleton so that everyone's file is consistent when they pull the repo. Added comments describing what each of the prototype methods will do once implemented.